+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/data/osmnotes/notequests/OsmNoteQuestAndroid.kt
@@ -8,12 +8,15 @@ import de.westnordost.streetcomplete.data.quest.Quest
 import de.westnordost.streetcomplete.data.quest.QuestType
 
 /** Represents one task for the user to contribute to a public OSM note */
-data class OsmNoteQuest(
-    val id: Long,
+data class OsmNoteQuestAndroid(
+    override val id: Long,
     override val position: LatLon
-) : Quest {
+) : OsmNoteQuest, Quest {
     override val type: QuestType get() = OsmNoteQuestType
     override val key: OsmNoteQuestKey by lazy { OsmNoteQuestKey(id) }
     override val markerLocations: Collection<LatLon> by lazy { listOf(position) }
     override val geometry: ElementGeometry get() = ElementPointGeometry(position)
 }
+
+actual fun createOsmNoteQuest(id: Long, position: LatLon): OsmNoteQuest =
+    OsmNoteQuestAndroid(id, position)
