--- app/src/androidMain/kotlin/de/westnordost/streetcomplete/data/osmnotes/notequests/OsmNoteQuestType.kt
@@ -1,17 +1,0 @@
-package de.westnordost.streetcomplete.data.osmnotes.notequests
-
-import de.westnordost.streetcomplete.R
-import de.westnordost.streetcomplete.data.quest.AndroidQuest
-import de.westnordost.streetcomplete.data.quest.QuestType
-import de.westnordost.streetcomplete.data.user.achievements.EditTypeAchievement
-import de.westnordost.streetcomplete.quests.note_discussion.NoteDiscussionForm
-
-object OsmNoteQuestType : QuestType, AndroidQuest {
-    override val icon = R.drawable.ic_quest_notes
-    override val title = R.string.quest_noteDiscussion_title
-    override val wikiLink = "Notes"
-    override val achievements = emptyList<EditTypeAchievement>()
-    override val visibilityEditable = false
-
-    override fun createForm() = NoteDiscussionForm()
-}
