+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/quests/note_discussion/OsmNoteQuestType.kt
@@ -1,0 +1,11 @@
+package de.westnordost.streetcomplete.quests.note_discussion
+
+object OsmNoteQuestType : QuestType, AndroidQuest {
+    override val icon = R.drawable.ic_quest_notes
+    override val title = R.string.quest_noteDiscussion_title
+    override val wikiLink = "Notes"
+    override val achievements = emptyList<EditTypeAchievement>()
+    override val visibilityEditable = false
+
+    override fun createForm() = NoteDiscussionForm()
+}
