+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/osm/osmquests/OsmQuestModule.kt
@@ -1,5 +1,6 @@
 package de.westnordost.streetcomplete.data.osm.osmquests
 
+import org.koin.core.qualifier.named
 import org.koin.dsl.module
 
 val osmQuestModule = module {
@@ -7,4 +8,5 @@ val osmQuestModule = module {
     factory { OsmQuestsHiddenDao(get()) }
 
     single<OsmQuestSource> { get<OsmQuestController>() }
+    single { OsmQuestController(get(), get(), get(), get(), get(named("CountryBoundariesLazy"))) }
 }
