+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/osmnotes/NotesModule.kt
@@ -1,13 +1,10 @@
 package de.westnordost.streetcomplete.data.osmnotes
 
-import android.content.Context
 import de.westnordost.streetcomplete.ApplicationConstants
-import kotlinx.io.files.Path
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
 
 val notesModule = module {
-    factory(named("AvatarsCacheDirectory")) { Path(get<Context>().cacheDir.path, ApplicationConstants.AVATARS_CACHE_DIRECTORY) }
     factory { AvatarsDownloader(get(), get(), get(), get(named("AvatarsCacheDirectory"))) }
     factory { AvatarsInNotesUpdater(get()) }
     factory { NoteDao(get()) }
