+++ app/src/nativeMain/kotlin/de/westnordost/streetcomplete/data/osm/osmquests/OsmQuestUtils.native.kt
@@ -1,0 +1,6 @@
+package de.westnordost.streetcomplete.data.osm.osmquests
+
+// TODO multiplatform: can be moved once all the quest types are in common
+actual fun getAnalyzePriority(questType: OsmElementQuestType<*>): Int {
+    TODO("Not yet implemented")
+}
