+++ app/src/nativeMain/kotlin/de/westnordost/streetcomplete/data/osmnotes/notequests/OsmNoteQuest.native.kt
@@ -1,0 +1,7 @@
+package de.westnordost.streetcomplete.data.osmnotes.notequests
+
+import de.westnordost.streetcomplete.data.osm.mapdata.LatLon
+
+actual fun createOsmNoteQuest(id: Long, position: LatLon): OsmNoteQuest {
+    TODO("Not yet implemented")
+}
